break: rename `worktop/durable` -> `worktop/cfw.durable`
