+++ packages/worktop/src/cfw.d.ts
@@ -2,7 +2,7 @@
 
 import type { Context, Initializer } from 'worktop';
 import type { Promisable, Strict } from 'worktop/utils';
-import type { Durable } from 'worktop/durable';
+import type { Durable } from 'worktop/cfw.durable';
 import type { KV } from 'worktop/kv';
 
 declare global {
