+++ packages/worktop/types/durable.ts
@@ -1,10 +1,10 @@
 import { Router } from 'worktop';
-import { Actor } from 'worktop/durable';
+import { Actor } from 'worktop/cfw.durable';
 import * as cookies from 'worktop/cookie';
 import * as utils from 'worktop/utils';
 
 import type { Bindings } from 'worktop/cfw';
-import type { Durable } from 'worktop/durable';
+import type { Durable } from 'worktop/cfw.durable';
 import type { WebSocket } from 'worktop/ws';
 import type { KV } from 'worktop/kv';
