+++ readme.md
@@ -156,12 +156,12 @@ The `worktop/kv` submodule contains all classes and utilities related to [Worker
 
 The `worktop/cache` submodule contains all utilities related to [Cloudflare's Cache](https://developers.cloudflare.com/workers/learning/how-the-cache-works).
 
-### Module: `worktop/durable`
+### Module: `worktop/cfw.durable`
 
-> [View `worktop/durable` API documentation](/src/durable.d.ts)
-<!-- > [View `worktop/durable` API documentation](/docs/module.durable.md) -->
+> [View `worktop/cfw.durable` API documentation](/src/cfw.durable.d.ts)
+<!-- > [View `worktop/cfw.durable` API documentation](/docs/module.cfw.durable.md) -->
 
-The `worktop/durable` submodule includes native types for [Durable Objects](https://developers.cloudflare.com/workers/runtime-apis/durable-objects) as well as an `Actor` abstract class that provides a blueprint for authoring a Durable Object that handles WebSocket connections.
+The `worktop/cfw.durable` submodule includes native types for Cloudflare's [Durable Objects](https://developers.cloudflare.com/workers/runtime-apis/durable-objects) as well as an `Actor` abstract class that provides a blueprint for authoring a Durable Object that handles WebSocket connections.
 
 > **Note:** Durable Objects can only be used with the Module Worker format. You must integrate the `Router` with the `worktop/modules` submodule.
